+++ woorepie/src/main/java/com/piehouse/woorepie/estate/entity/EstatePrice.java
@@ -26,7 +26,7 @@ public class EstatePrice {
     private Estate estate;
 
     @Column(nullable = false)
-    private Integer estatePrice;
+    private Long estatePrice;
 
     @CreationTimestamp
     @Column(updatable = false)
